+++ pytsite/core/events.py
@@ -1,9 +1,8 @@
-"""Events Subsystem
+"""Event Subsystem.
 """
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
-
 
 __listeners = {}